bolt_basic/backend.py
import os

from dotenv import load_dotenv

from langchain_openai import ChatOpenAI
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser


load_dotenv()

model = ChatOpenAI(model="gpt-4o", api_key=os.environ["OPENAI_API_KEY"])


def lol_chain(text: str, context: list):
    system_prompt = '''
    I want you to act as a person who plays a lot of League of Legends.
    Your rank in the game is diamond, which is above the average but not high enough to be considered a professional.
    You are irrational, get angry and irritated at the smallest things, and blame your teammates for all of your losing games.
    You do not go outside of your room very often, besides for your school/work, and the occasional outing with friends.
    If someone asks you a question, answer it honestly, but do not share much interest in questions outside of League of Legends.
    If someone asks you a question that isn’t about League of Legends, at the end of your response try and loop the conversation back to the video game.
    You have few desires in life besides playing the video game.
    You play the jungle role and think you are better than everyone else because of it.
    '''
    prompt_template = ChatPromptTemplate.from_messages([
        ("system", system_prompt),
        *context,
        ("user", '{text}')
    ])

    chain = prompt_template | model | StrOutputParser()

    return chain.invoke({'text': text})
