bolt_basic/requirements.txt
slack_bolt
python-dotenv
langchain
langchain-openai
ipython